Switch to POST

+++ src/main.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, request, redirect, url_for, send_file
+from flask import Flask, render_template, request, send_file
 from string import Template
 from datetime import date
 import re
@@ -16,39 +16,37 @@ with open(os.path.join(os.path.join(src_path, "templates"), 'template.tex'), 'r'
     template_file = f.read()
 
 
-@app.route("/")
+@app.route("/", methods=['GET', 'POST'])
 def main():
-    return render_template("index.html")
-
-
-@app.route("/generate")
-def generate():
-    args = dict(request.args)
-    try:
-        if not all(len(string) <= 64 for string in args.values()):
-            raise Exception("Invalid values entered")
-        args["fontdir"] = font_path
-        args["font"] = "osifont-lgpl3fe.ttf"
-        args["date"] = date.fromisoformat(args.get("date")).strftime("%d.%m.%Y")
-        if args.get("groupname").isalnum() and args.get("surname").isalpha() and bool(re.match('[\w\s]+$', args.get("docname"))) and bool(re.match('[\w\-]+$', args.get("id"))):
-            fill = Template(template_file)
-            result = fill.substitute(args)
-            with tempfile.NamedTemporaryFile() as tmp:
-                tmp_name = tmp.name
-                tmp.write(result.encode("utf-8"))
-                tmp.flush()
-                os.system(f"xelatex --interaction=nonstopmode --output-directory={os.path.dirname(tmp_name)} {tmp_name} >/dev/null")
-                with open(tmp.name + ".pdf", 'rb') as pdf_f:
-                    pdf = BytesIO(pdf_f.read())
-            for filename in glob.glob(tmp_name + "*"):
-                os.remove(filename)
-            return send_file(pdf, mimetype='application/pdf', download_name=f'TZO-{args.get("surname")}-{args.get("id")}.pdf')
-        else:
-            raise Exception("Invalid length or characters")
-        return redirect(url_for('main'))
-    except Exception as e:
-        print(e)
-        return redirect(url_for('main'))
+    if request.method == 'POST':
+        args = dict(request.form)
+        print(args)
+        try:
+            if not all(len(string) <= 64 for string in args.values()):
+                raise Exception("Invalid values entered")
+            args["fontdir"] = font_path
+            args["font"] = "osifont-lgpl3fe.ttf"
+            args["date"] = date.fromisoformat(args.get("date")).strftime("%d.%m.%Y")
+            if args.get("groupname").isalnum() and args.get("surname").isalpha() and bool(re.match('[\w\s]+$', args.get("docname"))) and bool(re.match('[\w\-]+$', args.get("id"))):
+                fill = Template(template_file)
+                result = fill.substitute(args)
+                with tempfile.NamedTemporaryFile() as tmp:
+                    tmp_name = tmp.name
+                    tmp.write(result.encode("utf-8"))
+                    tmp.flush()
+                    os.system(f"xelatex --interaction=nonstopmode --output-directory={os.path.dirname(tmp_name)} {tmp_name} >/dev/null")
+                    with open(tmp.name + ".pdf", 'rb') as pdf_f:
+                        pdf = BytesIO(pdf_f.read())
+                for filename in glob.glob(tmp_name + "*"):
+                    os.remove(filename)
+                return send_file(pdf, mimetype='application/pdf', download_name=f'TZO-{args.get("surname")}-{args.get("id")}.pdf')
+            else:
+                raise Exception("Invalid length or characters")
+        except Exception as e:
+            print(e)
+            return render_template("index.html", error=e)
+    else:
+        return render_template("index.html")
 
 
 @app.errorhandler(404)
